+++ src/icon.rs
@@ -90,7 +90,8 @@ impl Loaded {
     }
 
     fn from_png_path(path: impl AsRef<Path>) -> Result<Self> {
-        let decoder = png::Decoder::new(BufReader::new(std::fs::File::open(path)?));
+        let mut decoder = png::Decoder::new(BufReader::new(std::fs::File::open(path.as_ref())?));
+        decoder.set_transformations(png::Transformations::normalize_to_color8());
         let mut reader = decoder
             .read_info()
             .map_err(|e| anyhow!("failed to read png info: {}", e))?;
@@ -98,23 +99,14 @@ impl Loaded {
         let info = reader
             .next_frame(&mut buf)
             .map_err(|e| anyhow!("failed to read png frame: {}", e))?;
-
         let buf = &buf[..info.buffer_size()];
-        let info = reader.info();
 
         let data = match info.color_type {
             png::ColorType::Rgb => {
                 ensure!(buf.len() % 3 == 0, "corrupted icon file");
 
                 buf.chunks(3)
-                    .map(|chunk| {
-                        let a = 0xffu32 << 24;
-                        let r = u32::from(chunk[0]) << 16;
-                        let g = u32::from(chunk[1]) << 8;
-                        let b = u32::from(chunk[2]);
-
-                        a | r | g | b
-                    })
+                    .map(|chunk| u32::from_be_bytes([0xff, chunk[0], chunk[1], chunk[2]]))
                     .collect()
             }
             png::ColorType::Rgba => rgba_to_argb(buf)?,
@@ -122,49 +114,15 @@ impl Loaded {
                 ensure!(buf.len() % 2 == 0, "corrupted icon file");
 
                 buf.chunks(2)
-                    .map(|chunk| {
-                        let x = u32::from(chunk[0]);
-                        let a = u32::from(chunk[1]) << 24;
-
-                        a | (x << 16) | (x << 8) | x
-                    })
+                    .map(|chunk| u32::from_be_bytes([chunk[1], chunk[0], chunk[0], chunk[0]]))
                     .collect()
             }
             png::ColorType::Grayscale => buf
                 .iter()
                 .copied()
-                .map(u32::from)
-                .map(|chunk| {
-                    let a = 0xffu32 << 24;
-                    let r = chunk << 16;
-                    let g = chunk << 8;
-                    let b = chunk;
-
-                    a | r | g | b
-                })
+                .map(|pix| u32::from_be_bytes([0xff, pix, pix, pix]))
                 .collect(),
-            png::ColorType::Indexed => {
-                let palette = info.palette.as_ref().ok_or_else(|| {
-                    anyhow!("invalid image: palette is missing for indexed color type")
-                })?;
-
-                buf.iter()
-                    .copied()
-                    .map(|idx| {
-                        let start = 3 * usize::from(idx);
-                        let end = start + 3;
-                        let chunk = palette
-                            .get(start..end)
-                            .context("corrupted icon file (palette overflow)")?;
-                        let a = 0xffu32 << 24;
-                        let r = u32::from(chunk[0]) << 16;
-                        let g = u32::from(chunk[1]) << 8;
-                        let b = u32::from(chunk[2]);
-
-                        Ok(a | r | g | b)
-                    })
-                    .collect::<Result<_>>()?
-            }
+            png::ColorType::Indexed => unreachable!("image shall be converted"),
         };
 
         Ok(Self {